Split round.rs into a directory module
- Move round.rs to round/mod.rs
- Extract debug diagnostics into round/diagnostics.rs (cfg(debug_assertions))
- Extract test helper methods into round/test_helpers.rs (cfg(test))

This fixes an unused import warning in release builds caused by HandAnalyzer
being used only in debug code, and reduces the size of the main module.
